Only use FORCE_PROVISION for actually expensive tasks
Also some code reformatting

+++ source/software/software_flink.sh
@@ -10,17 +10,12 @@ source ./deploy_shared.sh
 source ./flink_shared.sh
 
 filename="flink-1.1.3"
-bindir="${FLINK_INSTALL_DIR}/bin"
 
-function downloadFlink {
-  filename="$filename.tgz"
-  echo "Flink: downloading..."
-  get_file "http://apache.mirrors.ovh.net/ftp.apache.org/dist/flink/flink-1.1.3/flink-1.1.3-bin-hadoop1-scala_2.10.tgz" $filename
-}
-
-echo "Setting up Flink"
-
-downloadFlink
-tar -xzf $filename -C /usr/local
-rm -rf $FLINK_INSTALL_DIR
-mv /usr/local/$(basename "$filename" .tgz) ${FLINK_INSTALL_DIR}
+# Install Flink
+if (($FORCE_INSTALL)) || ! [ -d $FLINK_INSTALL_DIR ]; then
+  echo "Flink: Installing..."
+  get_file "http://apache.mirrors.ovh.net/ftp.apache.org/dist/flink/flink-1.1.3/flink-1.1.3-bin-hadoop1-scala_2.10.tgz" "$filename.tgz"
+  tar xf "$filename.tgz"
+  rm -rf $FLINK_INSTALL_DIR
+  mv $filename $FLINK_INSTALL_DIR
+fi
